rename files, clean up code, add comments
