--- src/components/pages/course-form/CourseFormModal.jsx
@@ -1,20 +1,0 @@
-// import React from "react";
-// import { Modal } from "antd";
-
-// export default props => {
-//   const { visible, onCancel, courseToEdit, onSubmit } = props;
-
-//   return (
-//     <>
-//       <Modal
-//         title="Course Modal"
-//         visible={visible}
-//         onCancel={onCancel}
-//       >
-//         <CourseForm
-//           onSubmit={onSubmit}
-//         />
-//       </Modal>
-//     </>
-//   );
-// }
