+++ src/components/pages/form-module/index.js
@@ -1,5 +1,3 @@
-export { default as ModuleFormContainer } from './ModuleFormContainer';
 export { default as ModuleForm } from './ModuleForm';
-export { default as ModuleFormModal } from './ModuleFormModal';
 export { default as ModuleCard } from './ModuleCard';
 export { default as ModuleFormPage } from './ModuleFormPage';
