+++ src/components/pages/index.js
@@ -1,11 +1,11 @@
-export { CourseFormPage } from './course-form';
+export { CourseFormPage } from './form-course';
+export { ModuleForm, ModuleFormPage } from './form-module';
+export { ProgramForm } from './form-program';
 export { CourseView } from './course-view';
 export { HomePage } from './home';
 export { LoginPage } from './login';
-export { ModuleForm, ModuleFormPage } from './module-form';
 export { ModuleView } from './module-view';
 export { NotFoundPage } from './not-found';
-export { ProgramForm } from './program-form';
 export { ProgramView } from './program-view';
 export { SettingsPage } from './settings';
 export { ProfileEdit } from './settings';
